Move CommentTest and TypeOfItemTest into folder optionalparcelattributes.
